[FR] Update in forms.py
Two minor changes. At lines 220 and 251, the default value (if the submitting field is empty) is '' (empty string). The previous version resulted in a 'False' statement in the 'if'. Therefore, the value was changed, so that it returns a non-empty value (blank spaces). This therefore messed up with fields which were unrequired... It forced the cleaning of values we did not even write!